backend/tests/test_saas_middleware.py
# backend/tests/test_saas_middleware.py
"""
SaaS Multi-Tenancy Middleware Test Senaryoları

Bu testler manuel olarak çalıştırılabilir veya pytest ile otomatik test edilebilir.
"""

import pytest
from fastapi.testclient import TestClient
from datetime import datetime, timedelta
import sys
from pathlib import Path

# Backend path'i ekle
sys.path.insert(0, str(Path(__file__).parent.parent))

from app.main import app

client = TestClient(app)


class TestTenantStatusMiddleware:
    """Tenant durumu middleware testleri"""

    def test_suspended_tenant_blocked(self):
        """Suspended tenant erişimi engellenmelidir"""
        # 1. Super admin ile tenant suspend et
        # 2. Tenant kullanıcısı ile erişim dene
        # 3. 403 dönmesini bekle
        pass  # Manuel test için SQL script'i çalıştır

    def test_cancelled_tenant_blocked(self):
        """Cancelled tenant erişimi engellenmelidir"""
        pass

    def test_trial_expired_tenant_blocked(self):
        """Trial süresi dolmuş tenant erişimi engellenmelidir"""
        pass

    def test_active_tenant_allowed(self):
        """Active tenant erişimi izin vermelidir"""
        pass

    def test_superadmin_bypass(self):
        """Super admin tüm kontrolleri bypass etmelidir"""
        pass

    def test_public_endpoints_bypass(self):
        """Public endpoint'ler bypass edilmelidir"""
        # /health, /auth/token, /public/* gibi
        response = client.get("/health")
        assert response.status_code == 200


class TestSubscriptionLimitMiddleware:
    """Subscription limit middleware testleri"""

    def test_menu_limit_exceeded(self):
        """Menü item limiti aşımı engellenmelidir"""
        # 1. Basic plan (max 100 item)
        # 2. 100 item ekle
        # 3. 101. item eklemeyi dene
        # 4. 403 + LIMIT_EXCEEDED_MENU_ITEMS bekle
        pass

    def test_sube_limit_exceeded(self):
        """Şube limiti aşımı engellenmelidir"""
        pass

    def test_user_limit_exceeded(self):
        """Kullanıcı limiti aşımı engellenmelidir"""
        pass

    def test_read_operations_allowed(self):
        """GET işlemleri limit kontrolünden muaf olmalı"""
        # GET /menu/liste her zaman çalışmalı (limit dolmuş olsa bile)
        pass

    def test_enterprise_unlimited(self):
        """Enterprise plan limitsiz olmalı"""
        pass


class TestRowLevelSecurity:
    """PostgreSQL RLS testleri"""

    def test_tenant_isolation_menu(self):
        """Tenant A, Tenant B'nin menülerini görememeli"""
        # Direct SQL query ile test et
        pass

    def test_tenant_isolation_orders(self):
        """Tenant A, Tenant B'nin siparişlerini görememeli"""
        pass

    def test_superadmin_sees_all(self):
        """Super admin tüm tenant'ların verilerini görebilmeli"""
        pass


# ================================
# Manuel Test Script'leri
# ================================

"""
MANUEL TEST 1: Suspended Tenant
================================

-- PostgreSQL'de çalıştır:
UPDATE subscriptions SET status = 'suspended' WHERE isletme_id = 1;

-- API'de test et:
curl http://localhost:8000/menu/liste \
  -H "Authorization: Bearer TENANT_1_TOKEN" \
  -H "X-Sube-Id: 1"

-- Beklenen: 403 Forbidden
-- {
--   "ok": false,
--   "error_code": "SUBSCRIPTION_SUSPENDED",
--   "detail": "Aboneliğiniz askıya alınmış..."
-- }

-- Geri al:
UPDATE subscriptions SET status = 'active' WHERE isletme_id = 1;
"""

"""
MANUEL TEST 2: Menu Limit
=========================

-- 1. Tenant'ın mevcut menu sayısını kontrol et:
SELECT COUNT(*) FROM menu m
JOIN subeler s ON m.sube_id = s.id
WHERE s.isletme_id = 1;

-- 2. Subscription limitini mevcut sayıya eşitle:
UPDATE subscriptions
SET max_menu_items = (
  SELECT COUNT(*) FROM menu m
  JOIN subeler s ON m.sube_id = s.id
  WHERE s.isletme_id = 1
)
WHERE isletme_id = 1;

-- 3. Yeni menü eklemeyi dene:
curl -X POST http://localhost:8000/menu/ekle \
  -H "Authorization: Bearer TOKEN" \
  -H "X-Sube-Id: 1" \
  -H "Content-Type: application/json" \
  -d '{
    "ad": "Test Limit Ürün",
    "fiyat": 50.00,
    "kategori": "Test"
  }'

-- Beklenen: 403 Forbidden
-- {
--   "ok": false,
--   "error_code": "LIMIT_EXCEEDED_MENU_ITEMS",
--   "detail": "Menü item limiti aşıldı...",
--   "current": 100,
--   "limit": 100
-- }

-- Geri al:
UPDATE subscriptions SET max_menu_items = 100 WHERE isletme_id = 1;
"""

"""
MANUEL TEST 3: Trial Expired
============================

-- PostgreSQL'de çalıştır:
UPDATE subscriptions
SET status = 'trial',
    trial_baslangic = NOW() - INTERVAL '15 days',
    trial_bitis = NOW() - INTERVAL '1 day'
WHERE isletme_id = 1;

-- API'de test et:
curl http://localhost:8000/menu/liste \
  -H "Authorization: Bearer TOKEN" \
  -H "X-Sube-Id: 1"

-- Beklenen: 403 Forbidden
-- {
--   "ok": false,
--   "error_code": "TRIAL_EXPIRED",
--   "detail": "Deneme süreniz sona ermiş..."
-- }

-- Geri al:
UPDATE subscriptions
SET status = 'active',
    trial_baslangic = NULL,
    trial_bitis = NULL
WHERE isletme_id = 1;
"""

"""
MANUEL TEST 4: RLS Tenant Isolation
===================================

-- PostgreSQL'de çalıştır:

-- 1. İki farklı tenant oluştur
-- Tenant 1: isletme_id = 1, sube_id = 1
-- Tenant 2: isletme_id = 2, sube_id = 2

-- 2. Tenant 1 kullanıcısı olarak sorgu çalıştır
SET SESSION AUTHORIZATION 'tenant1_user';

-- 3. Tenant 2'nin şubesindeki menüleri görmeye çalış
SELECT * FROM menu WHERE sube_id = 2;

-- Beklenen: 0 rows (RLS engeller)

-- 4. Kendi şubesinin menülerini görebilir
SELECT * FROM menu WHERE sube_id = 1;

-- Beklenen: Kendi menüleri gelir

-- 5. Super admin bypass kontrolü
SET SESSION AUTHORIZATION 'superadmin_user';
SELECT * FROM menu WHERE sube_id = 2;

-- Beklenen: Tüm menüler gelir (bypass)
"""

"""
MANUEL TEST 5: Subscription Limits Endpoint
===========================================

-- Mevcut kullanımı görüntüle:
curl http://localhost:8000/subscription/1/limits \
  -H "Authorization: Bearer SUPER_ADMIN_TOKEN"

-- Beklenen response:
-- {
--   "plan_type": "basic",
--   "status": "active",
--   "limits": {
--     "max_subeler": 1,
--     "max_kullanicilar": 5,
--     "max_menu_items": 100
--   },
--   "usage": {
--     "subeler": 1,
--     "kullanicilar": 3,
--     "menu_items": 45
--   }
-- }
"""

"""
PERFORMANCE TEST: RLS Impact
=============================

-- RLS olmadan sorgu süresi:
EXPLAIN ANALYZE SELECT * FROM menu WHERE sube_id = 1;

-- RLS ile sorgu süresi:
ALTER TABLE menu ENABLE ROW LEVEL SECURITY;
EXPLAIN ANALYZE SELECT * FROM menu WHERE sube_id = 1;

-- Beklenen: Minimal performance impact (< 5%)
-- Eğer yavaşlama varsa index ekle:
CREATE INDEX idx_user_sube_izinleri_username
ON user_sube_izinleri(username);
"""


if __name__ == "__main__":
    print("SaaS Middleware Test Suite")
    print("=" * 50)
    print("\nBu testler manuel olarak çalıştırılmalıdır.")
    print("Yukarıdaki SQL script'lerini ve curl komutlarını kullanın.")
    print("\nOtomatik test için:")
    print("  pytest backend/tests/test_saas_middleware.py")
