backend/requirements.txt
aiosqlite==0.19.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
bcrypt==3.2.2
boto3==1.29.7
botocore==1.32.7
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comtypes==1.4.13
cryptography==46.0.3
databases==0.8.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.1
idna==3.11
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.1.0
pydantic_core==2.41.5
pydub==0.25.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyttsx3==2.90
pytz==2025.2
pywin32==311
PyYAML==6.0.3
redis==5.0.1
reportlab==4.4.4
requests==2.32.5
rsa==4.9.1
s3transfer==0.8.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.10.0
SQLAlchemy==1.4.54
starlette==0.41.3
structlog==23.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.0.7
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
